# optimized brigand::reverse