--- a/brigand/algorithms/reverse.hpp
+++ b/brigand/algorithms/reverse.hpp
@@ -6,8 +6,6 @@
 =================================================================================================**/
 #pragma once
 
-#include <brigand/types/integer.hpp>
-#include <brigand/sequences/size.hpp>
 #include <brigand/sequences/append.hpp>
 
 namespace brigand
@@ -15,75 +13,34 @@ namespace brigand
 namespace detail
 {
 
-    template <template <class...> class L, class First, class... R>
-    struct reverse_elements
+    template <template<class...> class L, class...>
+    struct reverse_elements;
+
+    template <template <class...> class L>
+    struct reverse_elements<L>
     {
-        using head = L<First>;
-        using tail = typename reverse_elements<L, R...>::type;
-        using type = typename brigand::append<tail, head>;
+        using type = L<>;
     };
 
-    template <template <class...> class L, class T0>
-    struct reverse_elements<L, T0>
+    template <template <class...> class L, class T0, class... Ts>
+    struct reverse_elements<L, T0, Ts...>
+    : append_impl<typename reverse_elements<L, Ts...>::type, L<T0>>
     {
-        using type = L<T0>;
     };
 
-    // fast lanes
-    template <template <class...> class L, class T0, class T1>
-    struct reverse_elements<L, T0, T1>
+    template <template <class...> class L, class T0, class T1, class T2, class T3, class T4, class T5, class T6, class T7, class T8, class T9, class T10, class T11, class T12, class T13, class T14, class T15, class... Ts>
+    struct reverse_elements<L, T0, T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, Ts...>
+    : append_impl<typename reverse_elements<L, Ts...>::type, L<T15, T14, T13, T12, T11, T10, T9, T8, T7, T6, T5, T4, T3, T2, T1, T0>>
     {
-        using type = L<T1, T0>;
     };
 
-    template <template <class...> class L, class T0, class T1, class T2>
-    struct reverse_elements<L, T0, T1, T2>
-    {
-        using type = L<T2, T1, T0>;
-    };
-
-    template <template <class...> class L, class T0, class T1, class T2, class T3>
-    struct reverse_elements<L, T0, T1, T2, T3>
-    {
-        using type = L<T3, T2, T1, T0>;
-    };
-
-    template <template <class...> class L, class T0, class T1, class T2, class T3, class T4>
-    struct reverse_elements<L, T0, T1, T2, T3, T4>
-    {
-        using type = L<T4, T3, T2, T1, T0>;
-    };
-
-    template <template <class...> class L, class T0, class T1, class T2, class T3, class T4, class T5>
-    struct reverse_elements<L, T0, T1, T2, T3, T4, T5>
-    {
-        using type = L<T5, T4, T3, T2, T1, T0>;
-    };
-
-    template <template <class...> class L, class T0, class T1, class T2, class T3, class T4, class T5, class T6>
-    struct reverse_elements<L, T0, T1, T2, T3, T4, T5, T6>
-    {
-        using type = L<T6, T5, T4, T3, T2, T1, T0>;
-    };
-
-    template <template <class...> class L, class T0, class T1, class T2, class T3, class T4, class T5, class T6, class T7>
-    struct reverse_elements<L, T0, T1, T2, T3, T4, T5, T6, T7>
-    {
-        using type = L<T7, T6, T5, T4, T3, T2, T1, T0>;
-    };
-
-    template <class L, class Count = brigand::size<L>> struct reverse_impl;
-
-    template<template<class...> class L, class... U, class Count>
-    struct reverse_impl<L<U...>,Count>
-    {
-        using type = typename reverse_elements<L, U...>::type;
-    };
+    template<class L>
+    struct reverse_impl;
 
     template<template<class...> class L, class... U>
-    struct reverse_impl<L<U...>, brigand::size_t<0>>
+    struct reverse_impl<L<U...>>
+    : reverse_elements<L, U...>
     {
-        using type = L<U...>;
     };
 
 }