Before:

Size= | 10 | 50 | 100 | 200 | 400
--- | --- | --- | --- | --- | ---
clang | 00.05s - 36M | 00.07s - 37M | 00.08s - 40M | 00.15s - 49M | ko
gcc | 00.04s - 19M | 00.05s - 21M | 00.07s - 25M | 00.13s - 38M | 00.32s - 80M

After:

Size= | 10 | 50 | 100 | 200 | 400
--- | --- | --- | --- | --- | ---
clang | 00.05s - 36M | 00.06s - 36M | 00.06s - 36M | 00.06s - 37M | 00.09s - 40M
gcc | 00.04s - 18M | 00.04s - 19M | 00.05s - 19M | 00.05s - 20M | 00.07s - 23M